+++ app/src/themes/theme_deletion_body.rs
@@ -1,25 +1,24 @@
-use crate::appearance::Appearance;
-use crate::send_telemetry_from_ctx;
-use crate::server::telemetry::TelemetryEvent;
-use crate::settings::{active_theme_kind, ThemeSettings};
-use crate::themes::theme::{ThemeKind, WarpTheme};
-use crate::user_config;
-use crate::user_config::util::from_yaml;
 use std::default::Default;
 use std::fs;
 use std::fs::remove_file;
+
 use warpui::assets::asset_cache::AssetSource;
 use warpui::elements::{
     Container, CornerRadius, CrossAxisAlignment, Flex, MainAxisSize, MouseStateHandle,
     ParentElement, Radius, SavePosition, Shrinkable, Text,
 };
 use warpui::fonts::Weight;
+use warpui::platform::Cursor;
 use warpui::ui_components::button::ButtonVariant;
 use warpui::ui_components::components::{Coords, UiComponent, UiComponentStyles};
-use warpui::{
-    platform::Cursor, AppContext, Element, Entity, SingletonEntity, TypedActionView, View,
-    ViewContext,
-};
+use warpui::{AppContext, Element, Entity, SingletonEntity, TypedActionView, View, ViewContext};
+
+use crate::appearance::Appearance;
+use crate::server::telemetry::TelemetryEvent;
+use crate::settings::{ThemeSettings, active_theme_kind};
+use crate::themes::theme::{ThemeKind, WarpTheme};
+use crate::user_config::util::from_yaml;
+use crate::{send_telemetry_from_ctx, user_config};
 
 const BUTTON_PADDING: f32 = 12.;
 const BUTTON_FONT_SIZE: f32 = 14.;
@@ -75,35 +74,38 @@ impl ThemeDeletionBody {
         let mut errored = true;
         let dir = user_config::themes_dir();
         // Check if the theme directory exists
-        if fs::metadata(&dir).is_ok() {
-            if let Some(ThemeKind::Custom(custom_theme)) = &self.theme_kind {
-                if let Ok(theme_from_yaml) = from_yaml::<WarpTheme>(custom_theme.path()) {
-                    // If theme has an image
-                    if let Some(image) = theme_from_yaml.background_image() {
-                        // Only delete the image if it is in the ./warp/themes directory.
-                        // We don't want to delete images from other parts of the user's filesystem.
-                        if let AssetSource::LocalFile { path } = image.source() {
-                            let image_path_in_themes_dir = dir.join(path.as_str());
-                            let _ = remove_file(image_path_in_themes_dir);
-                        } else {
-                            log::warn!("Attempted to delete a custom theme image with an unexpected image source");
-                        }
+        if fs::metadata(&dir).is_ok()
+            && let Some(ThemeKind::Custom(custom_theme)) = &self.theme_kind
+            && let Ok(theme_from_yaml) = from_yaml::<WarpTheme>(custom_theme.path())
+        {
+            // If theme has an image
+            if let Some(image) = theme_from_yaml.background_image() {
+                // Only delete the image if it is in the ./warp/themes directory.
+                // We don't want to delete images from other parts of the user's filesystem.
+                match image.source() {
+                    AssetSource::LocalFile { path, .. } => {
+                        let image_path_in_themes_dir = dir.join(path.as_str());
+                        let _ = remove_file(image_path_in_themes_dir);
                     }
-
-                    // Even if image can't be deleted, delete the theme .yaml file
-                    if remove_file(custom_theme.path()).is_ok() {
-                        let current_theme = active_theme_kind(ThemeSettings::as_ref(ctx), ctx);
-                        if matches!(current_theme, ThemeKind::Custom(theme) if &theme == custom_theme)
-                        {
-                            // Reset theme to Dark if we are deleting the current theme
-                            ctx.emit(ThemeDeletionBodyEvent::DeleteCurrentTheme)
-                        }
-                        errored = false;
-                        send_telemetry_from_ctx!(TelemetryEvent::DeleteCustomTheme, ctx);
-                        self.close(ctx);
-                        ctx.notify();
+                    _ => {
+                        log::warn!(
+                            "Attempted to delete a custom theme image with an unexpected image source"
+                        );
                     }
                 }
+            }
+
+            // Even if image can't be deleted, delete the theme .yaml file
+            if remove_file(custom_theme.path()).is_ok() {
+                let current_theme = active_theme_kind(ThemeSettings::as_ref(ctx), ctx);
+                if matches!(current_theme, ThemeKind::Custom(theme) if &theme == custom_theme) {
+                    // Reset theme to Dark if we are deleting the current theme
+                    ctx.emit(ThemeDeletionBodyEvent::DeleteCurrentTheme)
+                }
+                errored = false;
+                send_telemetry_from_ctx!(TelemetryEvent::DeleteCustomTheme, ctx);
+                self.close(ctx);
+                ctx.notify();
             }
         }
         if errored {
